Revert to using official Insecam sitemap as the 404 issue is resolved
This commit reverts previous changes where the project was adapted to use a local version of the Insecam sitemap due to a 404 error with the official URL. As the official sitemap URL is now functional again, this patch switches back to using it for sourcing camera feeds.

By reverting to the official sitemap, we ensure that we are always using the most current and updated data available.

The local version of the sitemap is also removed from the project configuration to avoid redundancy and potential confusion.